+++ packages/flet/lib/src/controls/container.dart
@@ -66,7 +66,6 @@ class ContainerControl extends StatelessWidget with FletStoreMixin {
         children.where((c) => c.name == "content" && c.isVisible);
     bool ink = control.attrBool("ink", false)!;
     bool onClick = control.attrBool("onclick", false)!;
-    bool onTap = control.attrBool("ontap", false)!;
     String url = control.attrString("url", "")!;
     String? urlTarget = control.attrString("urlTarget");
     bool onLongPress = control.attrBool("onLongPress", false)!;
@@ -146,7 +145,7 @@ class ContainerControl extends StatelessWidget with FletStoreMixin {
 
       Widget? result;
 
-      if ((onTap || onClick || url != "" || onLongPress || onHover) &&
+      if ((onClick || url != "" || onLongPress || onHover) &&
           ink &&
           !disabled) {
         var ink = Material(
@@ -156,12 +155,7 @@ class ContainerControl extends StatelessWidget with FletStoreMixin {
               // Dummy callback to enable widget
               // see https://github.com/flutter/flutter/issues/50116#issuecomment-582047374
               // and https://github.com/flutter/flutter/blob/eed80afe2c641fb14b82a22279d2d78c19661787/packages/flutter/lib/src/material/ink_well.dart#L1125-L1129
-              onTap: onTap
-                  ? () {
-                      debugPrint("Container ${control.id} Tap!");
-                      backend.triggerControlEvent(control.id, "tap", "");
-                    }
-                  : null,
+              onTap: onHover ? () {} : null,
               onTapDown: onClick || url != ""
                   ? (details) {
                       debugPrint("Container ${control.id} clicked!");
@@ -257,10 +251,10 @@ class ContainerControl extends StatelessWidget with FletStoreMixin {
                     : null,
                 child: child);
 
-        if ((onTap || onClick || onLongPress || onHover || url != "") &&
+        if ((onClick || onLongPress || onHover || url != "") &&
             !disabled) {
           result = MouseRegion(
-            cursor: onTap || onClick || url != ""
+            cursor: onClick || url != ""
                 ? SystemMouseCursors.click
                 : MouseCursor.defer,
             onEnter: onHover
@@ -278,12 +272,6 @@ class ContainerControl extends StatelessWidget with FletStoreMixin {
                   }
                 : null,
             child: GestureDetector(
-              onTap: onTap
-                  ? () {
-                      debugPrint("Container ${control.id} onTap!");
-                      backend.triggerControlEvent(control.id, "ontap", "");
-                    }
-                  : null,
               onTapDown: onClick || url != ""
                   ? (details) {
                       debugPrint("Container ${control.id} clicked!");