+++ sdk/python/packages/flet-core/src/flet_core/container.py
@@ -118,7 +118,6 @@ class Container(ConstrainedControl, AdaptiveControl):
         url_target: Optional[str] = None,
         theme: Optional[Theme] = None,
         theme_mode: Optional[ThemeMode] = None,
-        on_release=None,
         on_click=None,
         on_long_press=None,
         on_hover=None,
@@ -165,8 +164,6 @@ class Container(ConstrainedControl, AdaptiveControl):
             d = json.loads(e.data)
             return ContainerTapEvent(**d)
 
-        self.__on_release = EventHandler(convert_container_tap_event_data)
-        self._add_event_handler("tap", self.__on_release.get_handler())
         self.__on_click = EventHandler(convert_container_tap_event_data)
         self._add_event_handler("click", self.__on_click.get_handler())
 
@@ -195,7 +192,6 @@ class Container(ConstrainedControl, AdaptiveControl):
         self.url_target = url_target
         self.theme = theme
         self.theme_mode = theme_mode
-        self.on_release = on_release
         self.on_click = on_click
         self.on_long_press = on_long_press
         self.on_hover = on_hover
@@ -472,16 +468,6 @@ class Container(ConstrainedControl, AdaptiveControl):
         self.__theme_mode = value
         self._set_attr("themeMode", value.value if value is not None else None)
 
-    # on_release
-    @property
-    def on_release(self):
-        return self._get_event_handler("tap")
-
-    @on_release.setter
-    def on_release(self, handler):
-        self._add_event_handler("tap", handler)
-        self._set_attr("onTap", True if handler is not None else None)
-
     # on_click
     @property
     def on_click(self):
@@ -490,10 +476,8 @@ class Container(ConstrainedControl, AdaptiveControl):
     @on_click.setter
     def on_click(self, handler):
         self.__on_click.subscribe(handler)
-        if handler is not None:
-            self._set_attr("onclick", True)
-        else:
-            self._set_attr("onclick", None)
+        self._set_attr("onClick", True if handler is not None else None)
+
 
     # on_long_press
     @property