+++ sdk/python/packages/flet-core/src/flet_core/pubsub/pubsub_hub.py
@@ -7,7 +7,6 @@ from typing import Any, Awaitable, Callable, Dict, Iterable, Optional, Union
 import flet_core
 from flet_core.locks import NopeLock
 from flet_core.utils.concurrency_utils import is_pyodide
-from flet_core.utils.deprecated import deprecated
 
 logger = logging.getLogger(flet_core.__name__)
 
@@ -16,11 +15,11 @@ class PubSubHub:
     def __init__(
         self,
         loop: Optional[asyncio.AbstractEventLoop] = None,
-        pool: Optional[ThreadPoolExecutor] = None,
+        executor: Optional[ThreadPoolExecutor] = None,
     ):
         logger.debug("Creating new PubSubHub instance")
         self.__loop = loop
-        self.__pool = pool
+        self.__executor = executor
         self.__lock = threading.Lock() if not is_pyodide() else NopeLock()
         self.__subscribers: Dict[
             str, Union[Callable, Callable[..., Awaitable[Any]]]
@@ -137,114 +136,9 @@ class PubSubHub:
         if asyncio.iscoroutinefunction(handler):
             asyncio.run_coroutine_threadsafe(handler(*args), self.__loop)
         else:
-            if self.__pool:
+            if self.__executor:
                 self.__loop.call_soon_threadsafe(
-                    self.__loop.run_in_executor, self.__pool, handler, *args
+                    self.__loop.run_in_executor, self.__executor, handler, *args
                 )
             else:
                 handler(*args)
-
-
-class PubSub:
-    def __init__(self, pubsub: PubSubHub, session_id: str):
-        self.__pubsub = pubsub
-        self.__session_id = session_id
-
-    def send_all(self, message: Any):
-        self.__pubsub.send_all(message)
-
-    @deprecated(
-        reason="Use send_all() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def send_all_async(self, message: Any):
-        self.send_all(message)
-
-    def send_all_on_topic(self, topic: str, message: Any):
-        self.__pubsub.send_all_on_topic(topic, message)
-
-    @deprecated(
-        reason="Use send_all_on_topic() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def send_all_on_topic_async(self, topic: str, message: Any):
-        self.send_all_on_topic(topic, message)
-
-    def send_others(self, message: Any):
-        self.__pubsub.send_others(self.__session_id, message)
-
-    @deprecated(
-        reason="Use send_others() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def send_others_async(self, message: Any):
-        self.send_others(message)
-
-    def send_others_on_topic(self, topic: str, message: Any):
-        self.__pubsub.send_others_on_topic(self.__session_id, topic, message)
-
-    @deprecated(
-        reason="Use send_others_on_topic() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def send_others_on_topic_async(self, topic: str, message: Any):
-        self.send_others_on_topic(topic, message)
-
-    def subscribe(self, handler: Callable):
-        self.__pubsub.subscribe(self.__session_id, handler)
-
-    @deprecated(
-        reason="Use subscribe() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def subscribe_async(self, handler: Callable):
-        self.subscribe(handler)
-
-    def subscribe_topic(self, topic: str, handler: Callable):
-        self.__pubsub.subscribe_topic(self.__session_id, topic, handler)
-
-    @deprecated(
-        reason="Use subscribe_topic() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def subscribe_topic_async(self, topic: str, handler: Callable):
-        self.subscribe_topic(topic, handler)
-
-    def unsubscribe(self):
-        self.__pubsub.unsubscribe(self.__session_id)
-
-    @deprecated(
-        reason="Use unsubscribe() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def unsubscribe_async(self):
-        self.unsubscribe()
-
-    def unsubscribe_topic(self, topic: str):
-        self.__pubsub.unsubscribe_topic(self.__session_id, topic)
-
-    @deprecated(
-        reason="Use unsubscribe_topic() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def unsubscribe_topic_async(self, topic: str):
-        self.unsubscribe_topic(topic)
-
-    def unsubscribe_all(self):
-        self.__pubsub.unsubscribe_all(self.__session_id)
-
-    @deprecated(
-        reason="Use unsubscribe_all() method instead.",
-        version="0.21.0",
-        delete_version="1.0",
-    )
-    async def unsubscribe_all_async(self):
-        self.unsubscribe_all()