+++ packages/flet_audio_recorder/CHANGELOG.md
@@ -1,11 +1,19 @@
-## 0.20.2
+# 0.21.1
 
 No changes in this release. Version bumped to follow parent `flet` package.
 
-## 0.20.1
+# 0.21.0
 
 No changes in this release. Version bumped to follow parent `flet` package.
 
-## 0.20.0
+# 0.20.2
+
+No changes in this release. Version bumped to follow parent `flet` package.
+
+# 0.20.1
+
+No changes in this release. Version bumped to follow parent `flet` package.
+
+# 0.20.0
 
 Initial release of the package.